Add new random distribution (half)
Include a new random distribution  that returns a binary weighted
distribution so that there is a 50% chance the number will be in
the first "STEP", 25% in the second etc.

Also add a cache for the results of family_policy() as its called
a lot with the same settler ids and the value should be the same
(mostly) - an improvement should handle the sace where the direct
ancestor of two possible partners "ages out".